# specs/001-purchase-order-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Purchase Order API
  version: 1.0.0
paths:
  /purchase-orders:
    post:
      summary: Create a Purchase Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseOrderCreate'
      responses:
        '201':
          description: Created
    get:
      summary: Query Purchase Orders
      parameters:
        - in: query
          name: branch_id
        - in: query
          name: supplier_id
        - in: query
          name: status
      responses:
        '200':
          description: OK
  /purchase-orders/{poId}:
    get:
      summary: Get PO by ID
      parameters:
        - in: path
          name: poId
          required: true
      responses:
        '200':
          description: OK
  /purchase-orders/{poId}/fulfillment-history:
    get:
      summary: Get fulfillment history for a PO
      parameters:
        - in: path
          name: poId
          required: true
      responses:
        '200':
          description: OK
    post:
      summary: Submit a Draft PO
      parameters:
        - in: path
          name: poId
          required: true
      responses:
        '200':
          description: Submitted
  /purchase-orders/{poId}/approve:
    post:
      summary: Approve a Submitted PO
      parameters:
        - in: path
          name: poId
          required: true
      responses:
        '200':
          description: Approved
components:
  schemas:
    PurchaseOrderCreate:
      type: object
      properties:
        branch_id:
          type: string
        buyer_id:
          type: string
        supplier_id:
          type: string
        currency:
          type: string
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItemCreate'
    LineItemCreate:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        quantity:
          type: integer
        expected_price:
          type: number